Implement CSS into recipes page, change listed in body
Background is now blue, h1 tags are now bolded, h2 tags are underlined.
Home page is now center-aligned. Recipe pages are aligned left with a
clickable button that takes users home. The links are not decorated,
but their hover state is. There is a border around images as well.
Currently thinking of a way to make the descriptions pop out more.